requirements.txt
streamlit
pytesseract
pillow
chromadb
sentence-transformers
camel-tools
transformers
sentencepiece
torch
networkx
pandas
requests
beautifulsoup4
